// frontend/types/meditacao.ts
export type Meditacao = {
  id: number;
  data: string;
  titulo: string;
  subtitulo: string;
  leitura_ref: string;
  leitura_ref_pt: string;
  conteudo_i: string;
  conteudo_ii: string;
  conteudo_iii: string;
  titulo_pt: string;
  subtitulo_pt: string;
  conteudo_i_pt: string;
  conteudo_ii_pt: string;
  conteudo_iii_pt: string;
  fonte_traducao: string;
  criado_em: string;
};